\name{devx}
\alias{devx}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate Variance unsing C
}
\description{
Calculate variance of a numeric vector using C 
}
\usage{
devx(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A vector of numeric Values 
}
}
\details{
Calculate the variance using C loop
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
Input integer vector
}
\author{
Md.Shahjaman
}



