% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_one.R
\name{dist_one}
\alias{dist_one}
\title{Examine the distribution of a single continous variable}
\usage{
dist_one(data, ...)
}
\arguments{
\item{data}{The data frame or tibble}

\item{...}{The variable you would like to summarise
\itemize{
\item Default, with no additional arguments: density plot and histogram
\item Add the option "box" to show a boxplot and violin plot eg. data \%>\% dist_one(var, box)
\item Add the option "bee" to show a beeswarm plot eg. data \%>\% dist_one(var, bee)
}}
}
\value{
A tibble with a summary of the variable median, interquartile range, mean, SD, confidence intervals of the mean (using the Z distribution), and plots.
}
\description{
Summarises the median, interquartile range, mean, standard deviation, confidence intervals of the mean and plots the density and a histogram. Default option for a density plot and histogram can be overriden with the options "bee" and "box".

The function accepts an input from a dplyr pipe "\%>\%" and outputs the results as a tibble.
}
\examples{
example_data <- dplyr::tibble(id = 1:100, age = rnorm(100, mean = 30, sd = 10))
dist_one(example_data, age) # Default, with histogram and density plot
dist_one(example_data, age, box) # Option to show a boxplot and violin plot
dist_one(example_data, age, bee) # Option to show a beeswarm plot (using ggbeeswarm)
summary <- dist_one(example_data, age) # Save summary statistics as a tibble.
# Note that this function accepts a pipe input
# eg. example_data \%>\% dist_one(age)
}
