\name{sumR}
\alias{sumR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate the total of an integre x
} 
\description{
Calculate the total of an integre value x
}
\usage{
sumR(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A vector of numeric values
}
}
\details{
To Calculate sum of input vector using R
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
Input vector 
}
\author{
Md.Shahjaman
}


