% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup.R
\name{dup}
\alias{dup}
\title{Count the number of duplicates in a variable}
\usage{
dup(data, var)
}
\arguments{
\item{data}{The data frame or tibble}

\item{var}{The variable to assess}
}
\value{
A tibble with the number of duplicates found, and the number of missing values (NA), together with percentages.
Missing values are not considered as duplicates, and are excluded from the calculation of percentage of duplicate values.
}
\description{
Provides an integer value for the number of duplicates found wihtin a variable

The function accepts an input from a dplyr pipe "\%>\%" and outputs the results as a tibble.

eg. example_data \%>\% dup(variable)
}
\examples{
example_data <- dplyr::tibble(id = 1:200, age = round(rnorm(200, mean = 30, sd = 5), digits=0))
example_data$age[sample(1:200, size = 15)] <- NA  # Replace 15 values with missing.
dup(example_data, age)
# Note that this function accepts a pipe input
# eg. example_data \%>\% dup(group)
}
