% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumvar-package.R
\docType{package}
\name{sumvar-package}
\alias{sumvar}
\alias{sumvar-package}
\title{sumvar: Summarise Continuous and Categorical Variables, Check for Duplicates and Missing Data}
\description{
sumvar is a collection of functions useful for summarising variables. For continuous variables, the distribution of a single variable use dist_one() To explore a continuous variable stratified by a grouping variable use dist_sum() For a single categorical variable use tab1(), and a twoway table from two categorical variables- tab(). dist_sum() and tab() include options for hypothesis tests. Explore duplicates with dup()) and missing values with miss().
}
\author{
\strong{Maintainer}: Alexander Stockdale \email{a.stockdale@liverpool.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
